+++ chopper/test/multipart_test.dart
@@ -1,4 +1,5 @@
 import 'package:chopper/chopper.dart';
+import 'package:http_parser/http_parser.dart';
 import 'package:test/test.dart';
 import 'package:http/testing.dart';
 import 'package:http/http.dart' as http;
@@ -45,12 +46,13 @@ void main() {
           contains('content-type: application/octet-stream'),
         );
         expect(
+            req.body,
+            contains(
+              'content-disposition: form-data; name="file"',
+            ));
+        expect(
           req.body,
-          contains('''
-content-disposition: form-data; name="file"\r
-\r
-${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}\r
-'''),
+          contains('${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}'),
         );
         return http.Response('ok', 200);
       });
@@ -62,6 +64,40 @@ ${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}\r
 
       chopper.dispose();
     });
+  });
+
+  test('file with MultipartFile', () async {
+    final httpClient = MockClient((http.Request req) async {
+      expect(req.headers['Content-Type'], contains('multipart/form-data;'));
+      expect(
+        req.body,
+        contains('content-type: application/octet-stream'),
+      );
+      expect(
+          req.body,
+          contains(
+            'content-disposition: form-data; name="file_field"; filename="file_name"',
+          ));
+      expect(
+        req.body,
+        contains('${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}'),
+      );
+      return http.Response('ok', 200);
+    });
+
+    final chopper = ChopperClient(client: httpClient);
+    final service = HttpTestService.create(chopper);
+
+    final file = http.MultipartFile.fromBytes(
+      'file_field',
+      [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+      filename: 'file_name',
+      contentType: MediaType.parse('application/octet-stream'),
+    );
+
+    await service.postMultipartFile(file);
+
+    chopper.dispose();
   });
 
   test('PartValue', () async {
@@ -133,5 +169,43 @@ ${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}\r
     expect(req.files.length, equals(1));
     final bytes = await req.files.first.finalize().first;
     expect(bytes, equals([1, 2]));
+  });
+
+  test('PartValue with MultipartFile directly', () async {
+    final req = await toMultipartRequest(
+      [
+        PartValue<http.MultipartFile>(
+          null,
+          http.MultipartFile.fromBytes(
+            "first",
+            [1, 2],
+            filename: 'list int 1',
+          ),
+        ),
+        PartFile<http.MultipartFile>(
+          null,
+          http.MultipartFile.fromBytes(
+            "second",
+            [2, 1],
+            filename: 'list int 2',
+          ),
+        ),
+      ],
+      HttpMethod.Post,
+      Uri.parse('/foo'),
+      {},
+    );
+
+    final first = req.files[0];
+    final second = req.files[1];
+
+    expect(first.filename, equals('list int 1'));
+    expect(second.filename, equals('list int 2'));
+
+    var bytes = await first.finalize().first;
+    expect(bytes, equals([1, 2]));
+
+    bytes = await second.finalize().first;
+    expect(bytes, equals([2, 1]));
   });
 }